html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-optional-006.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: non-integer values for feature `left`</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var windowURL = 'resources/message-opener.html';
var featuresPrefix = `width=401,height=204,top=0,`;

// https://html.spec.whatwg.org/multipage/infrastructure.html#rules-for-parsing-integers
var baselineTop, baselineLeft;

setup (() => {
  var prefixedMessage = new PrefixedMessageTest();
  var featureString = `${featuresPrefix}left=0`;
  prefixedMessage.onMessage((data, e) => {
    e.source.close();
    baselineLeft = data.left;
    baselineTop = data.top;
    runWindowTests();
  });
  var win = window.open(prefixedMessage.url(windowURL), '', featureString);
});

function runWindowTests () {
  // When code point in first position is not an ASCII digit, "+" or "-",
  // that's an error and the value becomes 0
  [ 'left=/104',
    'left=_104',
    'left=L104',
    'left=0x68'
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.left, baselineLeft, `"${feature} begins with an invalid character and should be ignored"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL), '', featureString);
    }, `features "${feature}" should NOT set "left=104"`);
  });

  [ 'left=105.5',
    'left=105.32',
    'left=105LLl',
    'left=105^4',
    'left=105*3',
    'left=105/5',
    'left=105  '
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.left, 105, `"${featureString} value after first non-digit will be ignored"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL), '', featureString);
    }, `features "${feature}" should set "left=105"`);
  });
}

</script>
